Revert "configure.ac: don't depend on libosmogb."
gprs header files from libosmogb are needed for osmo-pcap. This was the
reason why it was added as dependency to configure.ac in
I9a8fa03cef1efc9fdaea65ee63ca9b3379993989.

  $ git grep "osmocom/gprs"
  src/osmo_client_core.c:#include <osmocom/gprs/gprs_bssgp.h>
  src/osmo_client_core.c:#include <osmocom/gprs/protocol/gsm_08_16.h>
  src/osmo_client_core.c:#include <osmocom/gprs/protocol/gsm_08_18.h>

The package is only still building for debian, because in debian
libosmocore is not split up, and for centos, because the rpm spec file
still lists the dependency.

This reverts commit 2c2eadcadf59ac42a02339f554c5daa48ee4f9fd.

Change-Id: Ic358956521230a2f6525533779c27417b2e86833