pcap-server: Introduce VTY cmd '[no] completed-path'
This VTY allows enabling feature to instruct osmo-pcap-server to move
the closed files (due to ration) into a separate directory.
This is useful for users who want to monitor and act on closed pcap
files to eg. generate statistics, etc.

Important: completed-path must be in the same filesystem mountpoint as
base-path, since osmo-pcap-server is actually not copying the files,
but atomically renaming the paths through rename() syscall.

Related: SYS#7248
Change-Id: I5166d1c5d9eb45358c87c2e1c5fbf7969d1d5294