pcap-server: Resolve real path of pcap file before opening
This means we always end up with a canonicalized absolute path.

While at it, also valdiate that we can indeed resolve the path (eg. it
exists and is reachable with current process permissions) during
startup.

It is preferrable to resolve it everytime a file is opened, this allows
the user to eg. change the base-path to a different symlink it they wish
to change the base-path for new pcaps without restarting osmo-pcap.

Change-Id: I8d161010dc8b480dd4cf90e19ca28a77914a50ad