server: Delay reopen of pcap only until necessary
If we are asked to reopen a pcap file (eg due to external SIGUSR1) while
in the middle of receiving a data packet (we already received the header
of the segment so data will arrive soon), code exists to delay reopening
so that we can include that last packet which arrived because the time
where we were asked to reopen.
However, the new pcap file was reopened only after next packet arrived.
Instead, we want to reopen it as soon as that last packet is received,
so that a new pcap file is created. This allows better tracking eg.
empty traffic during time in between last data packet before reopen and
the next data packet arriving.

While at it, rename the variable to make it more informative, and
convert it to a bool.

Change-Id: Id79500717a2e186aac979cded340a5af6ce3035b