server: Support storing of pcapng format
Improve existing logic doing validation on payload data and generalize
the file_hdr field to account for the possibily of receiving recordings
in pcapng format.

This requires, during LINK_HDR when file header is received, discerning
whether the recording is being done in pcap vs pcapng, (and in pcapng
figuring out the endianness of the file) and then applying validation
based on that knowledge.

With that knowledge, osmo-pcap-server can now store the pcap(ng) file
using the proper suffix for the file.

Related: SYS#5822
Change-Id: I275d28ab418a1514fa9c5c7c20f3d831cc6af8bb