server: Avoid lseek syscall on every packet write to pcap file
Instead of requesting the current size to the OS, keep track of the
amount of bytes stored in conn->wr_offset, and check agains that value
when we want to write new data to the file.
This should remove quite a lot of oeverhead when lots of traffic is
being handled.

Change-Id: I437eaca982d65d0f06e7d24863875f85267f0e44