client: Process up to 10 recorded pkts per poll iteration
This should decrease CPU use of lipcap capturing under high capturing
load, which is welcome now that we support osmo_io with io-uring backend
on the Tx side towards osmo-pcap-server.

The processing number is kept low in order to avoid starving the event
loop, as well as avoid filling up the msgb queue of the Tx side towards
osmo-pcap-server.

Related: SYS#7290
Change-Id: I569fc8489bae2b829ef4040d3c5111a61129de68